requirements.txt
azure-ai-translation-text
flask